Feat: 유저 상세 모달 페이지로 변경
